+++ src/app/demo/components-overview/lookup-examples/lookup-autocomplete-ac-example/lookup-autocomplete-ac-example.component.ts
@@ -25,7 +25,7 @@ export class LookupAutocompleteAcExampleComponent  extends LookupExampleComponen
     super.ngOnInit();
   }
 
-  start() {
+  reloadDataIntern() {
     this.lookupHandler.reloadData('normalautocomplete');
     this.lookupHandler.reloadData('reactiveautocomplete');
   }