+++ src/app/modules/lux-lookup/lux-lookup-combobox-ac/lux-lookup-combobox-ac.component.ts
@@ -73,6 +73,7 @@ export class LuxLookupComboboxAcComponent<T = LuxLookupTableEntry> extends LuxLo
   setLookupData(entries: LuxLookupTableEntry[]) {
     super.setLookupData(entries);
 
+    this.displayedEntries = [];
     this.updateDisplayedEntries();
   }